refactor: moved the test models to __test__/models.py
