+++ src/overzetten/__tests__/unit/config/test_model_naming.py
@@ -1,5 +1,5 @@
 from overzetten import DTO, DTOConfig
-from overzetten.__tests__.fixtures.sqlalchemy_models import User
+from overzetten.__tests__.fixtures.models import User
 
 
 def test_model_name_override():
