+++ src/overzetten/__tests__/unit/core/test_basic_dto_creation.py
@@ -4,7 +4,7 @@ import datetime
 from decimal import Decimal
 
 from overzetten import DTO, DTOConfig
-from overzetten.__tests__.fixtures.sqlalchemy_models import User
+from overzetten.__tests__.fixtures.models import User
 
 
 def test_basic_dto_creation():
